% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-utils.R
\name{arc_element_grob}
\alias{arc_element_grob}
\title{Create arc grob from theme element}
\usage{
arc_element_grob(element, ...)
}
\arguments{
\item{element}{Theme element, i.e. `element_rect` or similar.}

\item{...}{Other arguments to control specific of rendering. This is
usually at least position. See the source code for individual methods.}
}
\description{
Create arc grob from theme element
}
\keyword{internal}
