% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-utils.R
\name{arc_test}
\alias{arc_test}
\title{Test function}
\usage{
arc_test(plot = ggplot2::last_plot(), region = NULL, ...)
}
\arguments{
\item{plot}{a ggplot object.}

\item{region}{a `CELL` object.}

\item{...}{other parameters passing to `ArcPlot_build()`.}
}
\value{
return grob object invisibly.
}
\description{
Helper funtion used to quickly preview the results of the conversion.
}
\examples{
if (FALSE) { # Not Run
  library(ggplot2)
  library(grid)
  library(arcET)
  ggplot(mtcars, aes(wt, mpg)) + geom_point()
  arc_test()
}
}
\author{
Hou Yun
}
