% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon.R
\name{ArcAreaGrob}
\alias{ArcAreaGrob}
\title{Arc Area Grob}
\usage{
ArcAreaGrob(
  x = c(120, 90, 60),
  y = c(0.5, 0.8, 0.5),
  colour = "black",
  fill = "grey20",
  linewidth = 0.5,
  linetype = 1,
  alpha = NA,
  group = 1L,
  base_line = NULL,
  ...
)
}
\arguments{
\item{x}{numeric vector (in degree) specifying x-values.}

\item{y}{positive numeric vector (in radius) specifying y-values.}

\item{colour}{color of polygon.}

\item{fill}{background of polygon.}

\item{linewidth}{border size of polygon.}

\item{alpha}{transparency of polygon.}

\item{group}{numeric vector used to separate locations in x and y
into multiple polygons.}

\item{base_line}{positive numeric in [0, 1].}

\item{...}{other parameters passing to `ArcPolygonGrob()`.}
}
\value{
a grob object.
}
\description{
These functions can draw polygon based on path and baseline
on polar coordinate.
}
\author{
Hou Yun
}
