% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{ArcBezierGrob}
\alias{ArcBezierGrob}
\alias{ArcLinkGrob}
\title{Arc Link Grob}
\usage{
ArcBezierGrob(
  x = 20,
  y = 0.5,
  xend = 70,
  yend = 0.5,
  height = 0.6,
  n = 100,
  colour = "black",
  linewidth = 0.5,
  linetype = 1,
  alpha = NA,
  arrow = NULL,
  lineend = "butt",
  linejoin = "round",
  linemitre = 10,
  ...
)

ArcLinkGrob(
  x = 20,
  y = 0.5,
  xend = 70,
  yend = 0.7,
  arc = TRUE,
  colour = "black",
  linewidth = 0.5,
  linetype = 1,
  alpha = NA,
  ...
)
}
\arguments{
\item{x}{numeric vector (in degree) specifying x-values of start points.}

\item{y}{positive numeric vector (in radius) specifying y-values of start points.}

\item{xend}{numeric vector (in degree) specifying x-values of end points.}

\item{yend}{positive numeric vector (in radius) specifying y-values of end points.}

\item{height}{numeric value in [0, 1].}

\item{n}{integer value. The larger the value, the smoother the curve.}

\item{colour}{color of line.}

\item{linewidth}{line width in pt.}

\item{linetype}{line type, same as `lty` in `gpar()`.}

\item{alpha}{transparency of lines.}

\item{arrow}{NULL or arrow object created by `arrow()`.}

\item{lineend}{line end style (round, butt, square).}

\item{linejoin}{line join style (round, mitre, bevel).}

\item{linemitre}{line mitre limit (number greater than 1).}

\item{...}{not used.}
}
\value{
a grob object.
}
\description{
These functions can draw segments or bezier curve on polar coordinate.
}
\details{
`ArcLinkGrob()` used to drawing segment links, and `ArcBezierGrob()`
used to drawing bezier links on center.
}
\author{
Hou Yun
}
