% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{ArcChordGrob}
\alias{ArcChordGrob}
\title{Arc Chord Grob}
\usage{
ArcChordGrob(
  x = 20,
  y = 0.5,
  xend = 70,
  yend = 0.5,
  width = 10,
  width_end = 20,
  colour = NA,
  fill = "grey25",
  linewidth = 0.5,
  linetype = 1,
  alpha = NA,
  height = 0.6,
  n = 100,
  lineend = "butt",
  linejoin = "round",
  linemitre = 10,
  ...
)
}
\arguments{
\item{x}{numeric vector (in degree) specifying x-values of start points.}

\item{y}{positive numeric vector (in radius) specifying y-values of start points.}

\item{xend}{numeric vector (in degree) specifying x-values of end points.}

\item{yend}{positive numeric vector (in radius) specifying y-values of end points.}

\item{width}{width of start (in degree), should be less than 360.}

\item{width_end}{width of end (in degree), should be less than 360.}

\item{colour}{color of lines.}

\item{fill}{fill colour of lines}

\item{linewidth}{line width in pt.}

\item{linetype}{line type, same as `lty` in `gpar()`.}

\item{alpha}{transparency of lines.}

\item{height}{numeric value in [0, 1].}

\item{n}{integer value. The larger the value, the smoother the curve.}

\item{lineend}{line end style (round, butt, square).}

\item{linejoin}{line join style (round, mitre, bevel).}

\item{linemitre}{line mitre limit (number greater than 1).}

\item{...}{other parameters passing to `pieces_data()`.}
}
\value{
a grob object.
}
\description{
These functions can draw chord on polar coordinate.
}
\details{
`ArcLinkGrob()` used to drawing segment links, and `ArcBezierGrob()`
used to drawing bezier links on center.
}
\author{
Hou Yun
}
