% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggproto.R, R/layer-margintext.R,
%   R/layer-roundrect.R, R/layer-text.R
\docType{data}
\name{arcET-extensions}
\alias{arcET-extensions}
\alias{GeomMarginVtext}
\alias{GeomMarginHtext}
\alias{GeomRoundpolygon}
\alias{GeomRoundtile}
\alias{GeomRoundbar}
\alias{GeomRoundcol}
\alias{GeomRoundrect}
\alias{GeomBannerText}
\title{arcET package extensions to ggplot2}
\description{
In general the actual classes should be of little interest to users as the
standard ggplot2 api of using geom_* and stat_* functions for building up
the plot is encouraged.
}
\keyword{datasets}
\keyword{internal}
