% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line.R
\name{ArcHlineGrob}
\alias{ArcHlineGrob}
\alias{ArcVlineGrob}
\title{Arc Horizontal and Vertical Line}
\usage{
ArcHlineGrob(yintercept = 0.5, region = CELL(), ...)

ArcVlineGrob(xintercept = 90, region = CELL(), ...)
}
\arguments{
\item{yintercept}{intercept on the y axis.}

\item{region}{a CELL object (created by `CELL()` function) used to set
the panel area.}

\item{...}{other parameters passing to `ArcSegmentsGrob()`.}

\item{xintercept}{intercept on the x axis.}
}
\value{
a grob object.
}
\description{
`ArcHlineGrob()` can be used to draw horizontal line, and
`ArcVlineGrob()` can be used to draw vertical line.
}
\author{
Hou Yun
}
