% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margintext.R
\name{ArcMarginVtextGrob}
\alias{ArcMarginVtextGrob}
\alias{ArcMarginHtextGrob}
\title{Arc MarginText Grob}
\usage{
ArcMarginVtextGrob(
  label,
  x = NULL,
  y = 0.5,
  colour = "black",
  alpha = NA,
  size = 3.88,
  family = "",
  fontface = 1,
  lineheight = 1.2,
  parse = FALSE,
  sides = "l",
  margin = unit(1, "pt"),
  length = unit(1, "cm"),
  tick.length = "10\%",
  line.gp = gpar(),
  upper = NULL,
  lower = NULL,
  ...
)

ArcMarginHtextGrob(
  label,
  x = 90,
  y = NULL,
  colour = "black",
  alpha = NA,
  size = 3.88,
  family = "",
  fontface = 1,
  lineheight = 1.2,
  parse = FALSE,
  margin = unit(1, "pt"),
  length = unit(1, "cm"),
  tick.length = "10\%",
  line.gp = gpar(),
  upper = NULL,
  lower = NULL,
  ...
)
}
\arguments{
\item{label}{expression or character.}

\item{x}{numeric vector (in degree) specifying x-values.}

\item{y}{positive numeric vector (in radius) specifying y-values.}

\item{colour}{color of text.}

\item{alpha}{transparency of text.}

\item{size}{font size of text.}

\item{family}{font family of text.}

\item{fontface}{font face of text.}

\item{lineheight}{height of a line as a multiple of the size of text.}

\item{parse}{logical, if TRUE will parse text to expression.}

\item{sides}{position of margin text. `sides` must be one of "l" or "r" for
`ArcMarginVtextGrob()`, and must be one of "t" or "b" for `ArcMarginHtextGrob()`.}

\item{margin}{a grid unit object specifying top and bottom text margin.}

\item{length}{length of link, should be a grid unit object.}

\item{tick.length}{tick length of link, should be a grid unit object or character
ratio.}

\item{line.gp}{other parameters of link, see \code{?grid::gpar} for details.}

\item{upper, lower}{the upper and lower limits of allowable adjustments.}

\item{...}{not used.}
}
\value{
a grob object.
}
\description{
These functions can draw margin text on polar coordinate.
}
\author{
Hou Yun
}
