% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line.R
\name{ArcPathGrob}
\alias{ArcPathGrob}
\alias{ArcLineGrob}
\title{Arc Path Grob}
\usage{
ArcPathGrob(
  x = c(0, 20, 70),
  y = c(0.2, 0.5, 1),
  colour = "black",
  linewidth = 0.5,
  linetype = 1,
  alpha = NA,
  group = 1L,
  arrow = NULL,
  lineend = "butt",
  linejoin = "round",
  linemitre = 10,
  arc = TRUE,
  steps = 0.01,
  simplify = FALSE,
  ...
)

ArcLineGrob(...)
}
\arguments{
\item{x}{numeric vector (in degree) specifying x-values.}

\item{y}{positive numeric vector (in radius) specifying y-values.}

\item{colour}{color of line.}

\item{linewidth}{line width in pt.}

\item{linetype}{line type, same as `lty` in `gpar()`.}

\item{alpha}{transparency of lines.}

\item{group}{used to separate locations in x and y into sub-paths.}

\item{arrow}{arrow specification, as created by `grid::arrow()`.}

\item{lineend}{line end style (round, butt, square).}

\item{linejoin}{line join style (round, mitre, bevel).}

\item{linemitre}{line mitre limit (number greater than 1).}

\item{arc}{logical, if TRUE will split each path in small pieces.}

\item{steps}{step length used to split data. A smaller value means a
smoother curve.}

\item{simplify}{logical, When TRUE, line segments equal to x will not be split.}

\item{...}{not used.}
}
\value{
a grob object.
}
\description{
These functions can draw path on polar coordinate.
}
\author{
Hou Yun
}
