% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc.R
\name{arcplot}
\alias{arcplot}
\title{Create a new ArcPlot}
\usage{
arcplot(data = NULL, mapping = aes(), ...)
}
\arguments{
\item{data}{Default dataset to use for plot. If not already a data.frame,
will be converted to one by \code{\link[ggplot2:fortify]{fortify()}}. If not specified,
must be supplied in each layer added to the plot.}

\item{mapping}{Default list of aesthetic mappings to use for plot.
If not specified, must be supplied in each layer added to the plot.}

\item{...}{Other arguments passed on to methods. Not currently used.}
}
\value{
an ArcPlot object.
}
\description{
`arcplot()` initializes a ArcPlot object, and It works similarly
to `ggplot()`.
}
\author{
Hou Yun
}
