% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-build.R
\name{ArcPlot_build}
\alias{ArcPlot_build}
\alias{ArcPlot_build.ArcPlot}
\title{Build ArcPlot}
\usage{
ArcPlot_build(plot, ...)

\method{ArcPlot_build}{ArcPlot}(plot, ...)
}
\arguments{
\item{plot}{an ArcPlot object.}

\item{...}{other parameters passing to `merge_guides()`.}
}
\value{
grid grobs.
}
\description{
Internal function used to convert ArcPlot to grobs.
}
\author{
Hou Yun
}
