% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point.R
\name{ArcPointsGrob}
\alias{ArcPointsGrob}
\title{Arc Points Grob}
\usage{
ArcPointsGrob(
  x = 90,
  y = 0.5,
  shape = 1,
  size = 1,
  colour = "black",
  fill = NA,
  alpha = NA,
  stroke = 0.5,
  ...
)
}
\arguments{
\item{x}{numeric vector (in degree) specifying x-values.}

\item{y}{positive numeric vector (in radius) specifying y-values.}

\item{shape}{numeric or character vector specifying shape of symbols.}

\item{size}{numeric vector specifying size (in pt) of symbols.}

\item{colour}{color of symbols.}

\item{fill}{background of symbols when shape in 21-25.}

\item{alpha}{transparency of symbols.}

\item{stroke}{stroke size (in pt) of symbols.}

\item{...}{not used.}
}
\value{
a grob object.
}
\description{
These functions can draw symbols on polar coordinate.
}
\author{
Hou Yun
}
