% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon.R
\name{ArcPolygonGrob}
\alias{ArcPolygonGrob}
\title{Arc Polygon Grob}
\usage{
ArcPolygonGrob(
  x = c(30, 150, 150, 30),
  y = c(0.6, 0.6, 0.9, 0.9),
  colour = NA,
  fill = "grey20",
  linewidth = 0.5,
  linetype = 1,
  alpha = NA,
  group = 1L,
  subgroup = NULL,
  rule = "evenodd",
  lineend = "butt",
  linejoin = "round",
  linemitre = 10,
  arc = TRUE,
  steps = 0.005,
  simplify = TRUE,
  ...
)
}
\arguments{
\item{x}{numeric vector (in degree) specifying x-values.}

\item{y}{positive numeric vector (in radius) specifying y-values.}

\item{colour}{color of polygon.}

\item{fill}{background of polygon.}

\item{linewidth}{border size of polygon.}

\item{linetype}{border type of polygon.}

\item{alpha}{transparency of polygon.}

\item{group}{numeric vector used to separate locations in x and y
into multiple polygons.}

\item{subgroup}{just be used when polygons have holes.}

\item{rule}{character value specifying the fill rule: one of "winding" or
"evenodd" (default).}

\item{lineend}{line end style (round, butt, square).}

\item{linejoin}{line join style (round, mitre, bevel).}

\item{linemitre}{line mitre limit (number greater than 1).}

\item{arc}{logical, if TRUE will split each path in small pieces.}

\item{steps}{step length used to split data. A smaller value means a
smoother curve.}

\item{simplify}{logical, When TRUE, line segments equal to x will not be split.}

\item{...}{not used.}
}
\value{
a grob object.
}
\description{
These functions can draw polygon on polar coordinate.
}
\author{
Hou Yun
}
