% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\name{ArcRasterGrob}
\alias{ArcRasterGrob}
\title{Arc Raster Grob}
\usage{
ArcRasterGrob(
  image,
  x = 90,
  y = 0.5,
  fill = "white",
  angle = 0,
  size = 12,
  hjust = 0.5,
  vjust = 0.5,
  keep_asp = TRUE,
  read_args = list(),
  auto_adjust = TRUE,
  ...
)
}
\arguments{
\item{image}{raster object or others can be converted.}

\item{x}{numeric vector (in degree) specifying x-values.}

\item{y}{positive numeric vector (in radius) specifying y-values.}

\item{angle}{used to specifying angle of image, and should be one of
"downward", "reverse.clockwise", "clockwise", "downward", "inside" or
"outside".}

\item{size}{image size in "pt".}

\item{vjust}{numeric vector specifying vertical justification.}

\item{keep_asp}{logical. If TRUE will re-compute `height` and `width`
based on image dimmension.}

\item{...}{not used.}

\item{region}{a CELL object (created by `CELL()` function) used to set
the panel area.}

\item{clip}{logical, if TRUE will remove rows which overflow panel region.}

\item{steps}{step length used to split data. A smaller value means a
smoother curve.}

\item{simplify}{logical, When TRUE, line segments equal to x will not be split.}

\item{na.rm}{logical, if TRUE will remove NA's rows silently.}
}
\value{
a grob object.
}
\description{
These functions can draw raster image on polar coordinate.
}
\author{
Hou Yun
}
