% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon.R
\name{ArcRectGrob}
\alias{ArcRectGrob}
\alias{ArcPanelGrob}
\title{Arc Rect Grob}
\usage{
ArcRectGrob(
  xmin = 30,
  ymin = 0.6,
  xmax = 150,
  ymax = 0.9,
  colour = NA,
  fill = "grey20",
  linewidth = 0.5,
  linetype = 1,
  alpha = NA,
  ...
)

ArcPanelGrob(region = CELL(), fill = "grey85", colour = NA, ...)
}
\arguments{
\item{xmin, xmax}{numeric vector (in degree) specifying x-values of
start and end points.}

\item{ymin, ymax}{positive numeric vector (in radius) specifying y-values
of start and end points.}

\item{colour}{border of panel.}

\item{fill}{background of panel.}

\item{linewidth}{border size of rect.}

\item{linetype}{border type of rect.}

\item{alpha}{transparency of rect.}

\item{...}{other parameters passing to `ArcPolygonGrob()`.}

\item{region}{a CELL object (created by `CELL()` function) used to set
the drawing area.}
}
\value{
a grob object.
}
\description{
These functions can draw rect on polar coordinate.

Helper function to add background based on cell region.
}
\author{
Hou Yun
}
