% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon.R
\name{ArcRibbonGrob}
\alias{ArcRibbonGrob}
\title{Arc Ribbon Grob}
\usage{
ArcRibbonGrob(
  x = NULL,
  ymin = NULL,
  ymax = NULL,
  y = NULL,
  xmin = NULL,
  xmax = NULL,
  colour = NA,
  fill = "grey20",
  linewidth = 0.5,
  linetype = 1,
  alpha = NA,
  group = 1L,
  ...
)
}
\arguments{
\item{x}{numeric vector (in degree) specifying x-values.}

\item{ymin, ymax}{positive numeric vector (in radius) specifying bottom and top
border of horizontal ribbon.}

\item{y}{positive numeric vector (in radius) specifying y-values.}

\item{xmin, xmax}{numeric vector (in degree) specifying left and right border
of vertical ribbon.}

\item{colour}{color of ribbon.}

\item{fill}{background of ribbon.}

\item{linewidth}{border size of ribbon.}

\item{linetype}{border type of ribbon.}

\item{alpha}{transparency of ribbon.}

\item{group}{numeric vector used to separate locations in x and y
into multiple ribbons.}

\item{...}{other parameters passing to `ArcPolygonGrob()`.}
}
\value{
a grob object.
}
\description{
These functions can draw ribbon on polar coordinate.
}
\author{
Hou Yun
}
