% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtext.R
\name{ArcRichtextGrob}
\alias{ArcRichtextGrob}
\title{Arc Richtext Grob}
\usage{
ArcRichtextGrob(
  label,
  x = 90,
  y = 0.5,
  hjust = 0.5,
  vjust = 0.5,
  halign = hjust,
  valign = vjust,
  colour = "black",
  alpha = NA,
  angle = 0,
  size = 3.88,
  family = "",
  fontface = 1,
  lineheight = 1.2,
  margin = unit(c(0, 0, 0, 0), "pt"),
  padding = unit(c(0, 0, 0, 0), "pt"),
  r = unit(0, "pt"),
  align_widths = FALSE,
  align_heights = FALSE,
  box_gp = gpar(col = NA),
  use_markdown = TRUE,
  auto_adjust = TRUE,
  ...
)
}
\arguments{
\item{label}{character vector containing Markdown/HTML strings to draw.}

\item{x}{numeric vector (in degree) specifying x-values.}

\item{y}{positive numeric vector (in radius) specifying y-values.}

\item{hjust}{numeric vector specifying horizontal justification.}

\item{vjust}{numeric vector specifying vertical justification.}

\item{halign, valign}{Numerical values specifying the text justification
inside the text boxes. If not specified, these default to \code{hjust} and
\code{vjust}.}

\item{colour}{color of text.}

\item{alpha}{transparency of text.}

\item{angle}{angle of text.}

\item{size}{font size of text.}

\item{family}{font family of text.}

\item{fontface}{font face of text.}

\item{lineheight}{height of a line as a multiple of the size of text.}

\item{margin, padding}{Unit vectors of four elements each indicating the
margin and padding around each text label in the order top, right,
bottom, left. Margins are drawn outside the enclosing box (if any),
and padding is drawn inside. To avoid rendering artifacts, it is best
to specify these values in absolute units (such as points, mm, or inch)
rather than in relative units (such as npc).}

\item{r}{The radius of the rounded corners. To avoid rendering artifacts,
it is best to specify this in absolute units (such as points, mm, or inch)
rather than in relative units (such as npc).}

\item{align_widths, align_heights}{Should the widths and heights of all
the text boxes be aligned? Default is no.}

\item{box_gp}{Graphical parameters for the enclosing box around each text label.}

\item{use_markdown}{Should the \code{text} input be treated as markdown? Default
is yes.}

\item{auto_adjust}{logical, if TRUE will adjust angle based on x-axis value.}

\item{...}{not used.}
}
\value{
a grob object.
}
\description{
These functions can draw richtext on polar coordinate.
}
\author{
Hou Yun
}
