% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line.R
\name{ArcRugGrob}
\alias{ArcRugGrob}
\title{Arc Rug Grob}
\usage{
ArcRugGrob(
  x = NULL,
  y = NULL,
  colour = "black",
  linewidth = 0.5,
  linetype = 1,
  alpha = NA,
  sides = "bl",
  outside = FALSE,
  length = unit(2, "mm"),
  region = CELL(),
  lineend = "butt",
  clip = FALSE,
  ...
)
}
\arguments{
\item{x}{numeric vector (in degree) specifying x-values.}

\item{y}{positive numeric vector (in radius) specifying y-values.}

\item{colour}{color of lines.}

\item{linewidth}{line width in pt.}

\item{linetype}{line type, same as `lty` in `gpar()`.}

\item{alpha}{transparency of lines.}

\item{sides}{A string that controls which sides of the plot the rugs appear on.
It can be set to a string containing any of \code{"trbl"}, for top, right,
bottom, and left.}

\item{outside}{logical that controls whether to move the rug tassels outside of the plot area. Default is off (FALSE). You will also need to use \code{coord_cartesian(clip = "off")}. When set to TRUE, also consider changing the sides argument to "tr". See examples.}

\item{length}{A \code{\link[grid:unit]{grid::unit()}} object that sets the length of the rug lines. Use scale expansion to avoid overplotting of data.}

\item{region}{a CELL object (created by `CELL()` function) used to set
the panel area.}

\item{...}{not used.}
}
\value{
a grob object.
}
\description{
These functions can draw rugs on polar coordinate.
}
\author{
Hou Yun
}
