% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line.R
\name{ArcSegmentsGrob}
\alias{ArcSegmentsGrob}
\title{Arc Path Grob}
\usage{
ArcSegmentsGrob(
  x = 30,
  y = 0.4,
  xend = 150,
  yend = 0.4,
  colour = "black",
  linewidth = 0.5,
  linetype = 1,
  alpha = NA,
  arrow = NULL,
  lineend = "butt",
  linejoin = "round",
  linemitre = 10,
  arc = TRUE,
  steps = 0.01,
  simplify = FALSE,
  ...
)
}
\arguments{
\item{x}{numeric vector (in degree) specifying x-values of start point.}

\item{y}{positive numeric vector (in radius) specifying y-values of start point.}

\item{xend}{numeric vector (in degree) specifying x-values of end point.}

\item{yend}{positive numeric vector (in radius) specifying y-values of end point.}

\item{colour}{color of lines.}

\item{linewidth}{line width in pt.}

\item{linetype}{line type, same as `lty` in `gpar()`.}

\item{alpha}{transparency of lines.}

\item{arrow}{arrow specification, as created by `grid::arrow()`.}

\item{lineend}{line end style (round, butt, square).}

\item{linejoin}{line join style (round, mitre, bevel).}

\item{linemitre}{line mitre limit (number greater than 1).}

\item{arc}{logical, if TRUE will split each path in small pieces.}

\item{steps}{step length used to split data. A smaller value means a
smoother curve.}

\item{simplify}{logical, When TRUE, line segments equal to x will not be split.}

\item{...}{not used.}
}
\value{
a grob object.
}
\description{
These functions can draw segments on polar coordinate.
}
\author{
Hou Yun
}
