% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{ArcShapeGrob}
\alias{ArcShapeGrob}
\title{Arc Shape Grob}
\usage{
ArcShapeGrob(
  x = c(120, 60, 60, 120, 120),
  y = c(0.5, 0.5, 0.9, 0.9, 0.5),
  colour = NA,
  fill = "grey20",
  linewidth = 0.5,
  linetype = 1,
  alpha = NA,
  group = 1L,
  expand = 0,
  radius = 0,
  lineend = "butt",
  linejoin = "round",
  linemitre = 10,
  arc = TRUE,
  steps = 0.01,
  simplify = FALSE,
  ...
)
}
\arguments{
\item{x}{numeric vector (in degree) specifying x-values.}

\item{y}{positive numeric vector (in radius) specifying y-values.}

\item{colour}{color of shape.}

\item{fill}{background of shape.}

\item{linewidth}{border size of shape.}

\item{linetype}{border type of shape.}

\item{alpha}{transparency of shape.}

\item{group}{numeric vector used to separate locations in x and y
into multiple polygons.}

\item{expand}{A numeric or unit vector of length one, specifying the
expansion amount. Negative values will result in contraction instead. If the
value is given as a numeric it will be understood as a proportion of the
plot area width.}

\item{radius}{As \code{expand} but specifying the corner radius.}

\item{lineend}{line end style (round, butt, square).}

\item{linejoin}{line join style (round, mitre, bevel).}

\item{linemitre}{line mitre limit (number greater than 1).}

\item{arc}{logical, if TRUE will split each path in small pieces.}

\item{steps}{step length used to split data. A smaller value means a
smoother curve.}

\item{simplify}{logical, When TRUE, line segments equal to x will not be split.}

\item{...}{not used.}
}
\value{
a grob object.
}
\description{
These functions can draw shape on polar coordinate.
}
\author{
Hou Yun
}
