% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line.R
\name{ArcStepsGrob}
\alias{ArcStepsGrob}
\title{Arc Steps Grob}
\usage{
ArcStepsGrob(
  x = c(0, 20, 70),
  y = c(0.2, 0.5, 1),
  colour = "black",
  linewidth = 0.5,
  linetype = 1,
  alpha = NA,
  group = 1L,
  direction = "hv",
  ...
)
}
\arguments{
\item{x}{numeric vector (in degree) specifying x-values.}

\item{y}{positive numeric vector (in radius) specifying y-values.}

\item{colour}{color of lines.}

\item{linewidth}{line width in pt.}

\item{linetype}{line type, same as `lty` in `gpar()`.}

\item{alpha}{transparency of lines.}

\item{group}{used to separate locations in x and y into sub-paths.}

\item{direction}{direction of stairs: 'vh' for vertical then horizontal,
'hv' for horizontal then vertical, or 'mid' for step half-way between
adjacent x-values.}

\item{...}{other parameters passing to `ArcPathGrob()`.}
}
\value{
a grob object.
}
\description{
These functions can draw steps on polar coordinate.
}
\author{
Hou Yun
}
