% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{ArcTextGrob}
\alias{ArcTextGrob}
\title{Arc Text Grob}
\usage{
ArcTextGrob(
  label,
  x = 90,
  y = 0.5,
  hjust = 0.5,
  vjust = 0.5,
  colour = "black",
  alpha = NA,
  angle = 0,
  size = 3.88,
  family = "",
  fontface = 1,
  lineheight = 1.2,
  check_overlap = FALSE,
  parse = FALSE,
  auto_adjust = TRUE,
  ...
)
}
\arguments{
\item{label}{expression or character.}

\item{x}{numeric vector (in degree) specifying x-values.}

\item{y}{positive numeric vector (in radius) specifying y-values.}

\item{hjust}{numeric vector specifying horizontal justification.}

\item{vjust}{numeric vector specifying vertical justification.}

\item{colour}{color of text.}

\item{alpha}{transparency of text.}

\item{angle}{angle of text.}

\item{size}{font size of text.}

\item{family}{font family of text.}

\item{fontface}{font face of text.}

\item{lineheight}{height of a line as a multiple of the size of text.}

\item{check_overlap}{logical value to indicate whether to check for and omit
overlapping text.}

\item{parse}{logical, if TRUE will parse text to expression.}

\item{auto_adjust}{logical, if TRUE will adjust angle based on x-axis value.}

\item{...}{not used.}
}
\value{
a grob object.
}
\description{
These functions can draw text on polar coordinate.
}
\author{
Hou Yun
}
