% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon.R
\name{ArcTileGrob}
\alias{ArcTileGrob}
\title{Arc Tile Grob}
\usage{
ArcTileGrob(
  x = 90,
  y = 0.8,
  width = NULL,
  height = NULL,
  colour = NA,
  fill = "grey20",
  linewidth = 0.5,
  linetype = 1,
  alpha = NA,
  ...
)
}
\arguments{
\item{x}{numeric vector (in degree) specifying x-values.}

\item{y}{positive numeric vector (in radius) specifying y-values.}

\item{width}{one-length numeric specifying tile width.}

\item{height}{one-length numeric specifying tile height.}

\item{colour}{color of tile.}

\item{fill}{background of tile.}

\item{linewidth}{border size of tile.}

\item{alpha}{transparency of rect.}

\item{...}{other parameters passing to `ArcPolygonGrob()`.}
}
\value{
a grob object.
}
\description{
These functions can draw tile on polar coordinate.
}
\author{
Hou Yun
}
