% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{ArcyAxisGrob}
\alias{ArcyAxisGrob}
\title{Arc y-axis Grob}
\usage{
ArcyAxisGrob(
  coord = PANEL(),
  region = CELL(),
  position = "left",
  ticks.length = 1.5,
  line.gp = element_line(),
  tick.gp = element_line(),
  text.gp = element_text(),
  ...
)
}
\arguments{
\item{coord}{coordinate specification, as created by `PANEL()` or extract from
ggplot object.}

\item{region}{a CELL object (created by `CELL()` function) used to set
the drawing area.}

\item{position}{axis position specification, and default is "left".}

\item{ticks.length}{a grid unit object or numeric, and numerical value
means that the length of ticks is measured in millimeters.}

\item{line.gp}{should be created by `ggplot2::element_line()`.}

\item{tick.gp}{should be created by `ggplot2::element_line()`.}

\item{text.gp}{should be created by `ggplot2::element_text()`.}

\item{...}{not used.}
}
\value{
a `yAxisGrob` object.
}
\description{
These functions can draw y-axis on polar coordinate.
}
\author{
Hou Yun
}
