% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{CELL}
\alias{CELL}
\title{Cell Region}
\usage{
CELL(start = 360, end = 0, r0 = 0, r1 = 1, direct = "clockwise", zero = FALSE)
}
\arguments{
\item{start}{panel start point in degree, should be in [0, 360].}

\item{end}{panel end point in degree, should be in [0, 360].}

\item{r0}{panel start point in radius, and should be in [0, 1].}

\item{r1}{panel end point in radius, and should be in [0, 1].}

\item{direct}{must be one of "clockwise" or "reverse.clockwise". If "clockwise",
x-axis is clockwise.}

\item{zero}{logical. When `zero` is FALSE, `start` equals to `end` meaning
that the X-axis spans the entire circle.}
}
\value{
a `CELL` object.
}
\description{
This function always used to setting panel region on
polar coordinate. Note that the start point needs to be larger than
the end point.
}
\author{
Hou Yun
}
