% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-utils.R
\name{extract_ggplot}
\alias{extract_ggplot}
\alias{extract_params}
\alias{extract_params.default}
\alias{extract_guides}
\alias{extract_guides.default}
\title{Extract ggplot elements}
\usage{
extract_ggplot(plot)

extract_params(layer, ...)

\method{extract_params}{default}(layer, ...)

extract_guides(plot, ...)

\method{extract_guides}{default}(plot, ...)
}
\arguments{
\item{plot}{a ggplot object.}

\item{layer}{layers of ggplot object.}
}
\value{
elements list of plot.
}
\description{
Internal function to extract ggplot elements.
}
\author{
Hou Yun
}
