% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer-margintext.R
\name{geom_margin_vtext}
\alias{geom_margin_vtext}
\alias{geom_margin_htext}
\title{Margin Text}
\usage{
geom_margin_vtext(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  align = TRUE,
  parse = FALSE,
  margin = unit(1, "pt"),
  sides = "r",
  length = unit(1, "cm"),
  tick.length = unit(1.5, "mm"),
  line_colour = NULL,
  line_alpha = NULL,
  na.rm = FALSE,
  show.legend = FALSE,
  inherit.aes = TRUE
)

geom_margin_htext(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  align = TRUE,
  parse = FALSE,
  margin = unit(1, "pt"),
  sides = "t",
  length = unit(1, "cm"),
  tick.length = unit(1.5, "mm"),
  line_colour = NULL,
  line_alpha = NULL,
  na.rm = FALSE,
  show.legend = FALSE,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, either as a \code{ggproto} \code{Geom} subclass or as a string naming the
stat stripped of the \code{stat_} prefix (e.g. \code{"count"} rather than
\code{"stat_count"})}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function. Cannot be jointly specified with
\code{nudge_x} or \code{nudge_y}.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{align}{logical indicating wheather align label.}

\item{parse}{logical, if TRUE will parse text to expression.}

\item{margin}{a grid unit object specifying top and bottom text margin.}

\item{sides}{position of margin text. `sides` must be one of "l" or "r" for
`MarginVtextGrob()`, and must be one of "t" or "b" for `MarginHtextGrob()`.}

\item{length}{length of link, should be a grid unit object.}

\item{tick.length}{tick length of link, should be a grid unit object or character
ratio.}

\item{line_colour, line_alpha}{aesthetic of link lines.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\value{
a ggplot layer object.
}
\description{
The margin text geom is used to create margin text. In contrast
to `geom_text()`, `geom_margin_*text()` automatically adjusts its position
based on the graph size to avoid overlap.
}
\author{
Hou Yun
}
