% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer-blank.R
\name{GeomBlank2grob}
\alias{GeomBlank2grob}
\title{Convert Layer to Grob}
\usage{
GeomBlank2grob(data, trans = NULL, coord = PANEL(), region = CELL(), ...)
}
\arguments{
\item{data}{data frame object, which is extract from a ggplot object.}

\item{trans}{coordinate transform function.}

\item{coord}{coordinate specification, as created by `PANEL()` or extract from
ggplot object.}

\item{region}{a CELL object (created by `CELL()` function) used to set
the drawing area.}

\item{...}{other parameters passing to `Arc*Grob()` function.}
}
\value{
a grid grob object.
}
\description{
Convert a ggplot layer to arc grob.
}
\seealso{
Other transform: 
\code{\link{GeomBar2grob}()},
\code{\link{GeomBoxplot2grob}()},
\code{\link{GeomLinerange2grob}()},
\code{\link{GeomMarginVtext2grob}()},
\code{\link{GeomPath2grob}()},
\code{\link{GeomPoint2grob}()},
\code{\link{GeomPolygon2grob}()},
\code{\link{GeomRichText2grob}()},
\code{\link{GeomRoundcol2grob}()},
\code{\link{GeomText2grob}()},
\code{\link{GeomViolin2grob}()}
}
\author{
Hou Yun
}
\concept{transform}
