% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer-boxplot.R
\name{GeomBoxplot2grob}
\alias{GeomBoxplot2grob}
\title{Convert Layer to Grob}
\usage{
GeomBoxplot2grob(
  data,
  trans = NULL,
  coord = PANEL(),
  region = CELL(),
  ...,
  flipped_aes = FALSE,
  notch = FALSE,
  notchwidth = 0.5,
  fatten = 2,
  outlier.colour = NULL,
  outlier.color = NULL,
  outlier.fill = NULL,
  outlier.shape = 19,
  outlier.size = 1.5,
  outlier.stroke = 0.5,
  outlier.alpha = NULL,
  clip = FALSE,
  na.rm = FALSE
)
}
\arguments{
\item{data}{data frame object, which is extract from a ggplot object.}

\item{trans}{coordinate transform function.}

\item{coord}{coordinate specification, as created by `PANEL()` or extract from
ggplot object.}

\item{region}{a CELL object (created by `CELL()` function) used to set
the drawing area.}

\item{...}{other parameters passing to `Arc*Grob()` function.}

\item{flipped_aes}{TRUE means that coordinates are inherit `CoordFlip`.}

\item{notch}{If \code{FALSE} (default) make a standard box plot. If
\code{TRUE}, make a notched box plot. Notches are used to compare groups;
if the notches of two boxes do not overlap, this suggests that the medians
are significantly different.}

\item{notchwidth}{For a notched box plot, width of the notch relative to
the body (defaults to \code{notchwidth = 0.5}).}

\item{fatten}{A multiplicative factor used to increase the size of the
middle bar in \code{geom_crossbar()} and the middle point in
\code{geom_pointrange()}.}

\item{outlier.colour, outlier.color, outlier.fill, outlier.shape, outlier.size, outlier.stroke, outlier.alpha}{default
aesthetics for outliers. Set to NULL to inherit from the aesthetics used for the box.}

\item{clip}{logical. Allows points to overflow outside the drawing area when
`clip` is FALSE.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}
}
\value{
a grid grob object.
}
\description{
Convert a ggplot layer to arc grob.
}
\seealso{
Other transform: 
\code{\link{GeomBar2grob}()},
\code{\link{GeomBlank2grob}()},
\code{\link{GeomLinerange2grob}()},
\code{\link{GeomMarginVtext2grob}()},
\code{\link{GeomPath2grob}()},
\code{\link{GeomPoint2grob}()},
\code{\link{GeomPolygon2grob}()},
\code{\link{GeomRoundcol2grob}()},
\code{\link{GeomText2grob}()},
\code{\link{GeomViolin2grob}()}
}
\author{
Hou Yun
}
\concept{transform}
