% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer-range.R
\name{GeomLinerange2grob}
\alias{GeomLinerange2grob}
\alias{GeomPointrange2grob}
\alias{GeomErrorbar2grob}
\alias{GeomErrorbarh2grob}
\alias{GeomCrossbar2grob}
\title{Convert Layer to Grob}
\usage{
GeomLinerange2grob(
  data,
  trans = NULL,
  coord = PANEL(),
  region = CELL(),
  ...,
  flipped_aes = FALSE,
  clip = FALSE,
  na.rm = FALSE
)

GeomPointrange2grob(
  data,
  trans = NULL,
  coord = PANEL(),
  region = CELL(),
  ...,
  flipped_aes = FALSE,
  fatten = 4,
  clip = FALSE,
  na.rm = FALSE
)

GeomErrorbar2grob(
  data,
  trans = NULL,
  coord = PANEL(),
  region = CELL(),
  ...,
  flipped_aes = FALSE,
  clip = FALSE,
  na.rm = FALSE
)

GeomErrorbarh2grob(
  data,
  trans = NULL,
  coord = PANEL(),
  region = CELL(),
  ...,
  flipped_aes = FALSE,
  clip = FALSE,
  na.rm = FALSE
)

GeomCrossbar2grob(
  data,
  trans = NULL,
  coord = PANEL(),
  region = CELL(),
  ...,
  flipped_aes = FALSE,
  fatten = 2.5,
  clip = FALSE,
  na.rm = FALSE
)
}
\arguments{
\item{data}{data frame object, which is extract from a ggplot object.}

\item{trans}{coordinate transform function.}

\item{coord}{coordinate specification, as created by `PANEL()` or extract from
ggplot object.}

\item{region}{a CELL object (created by `CELL()` function) used to set
the drawing area.}

\item{...}{other parameters passing to `Arc*Grob()` function.}

\item{flipped_aes}{TRUE means that coordinates are inherit `CoordFlip`.}

\item{clip}{logical. Allows points to overflow outside the drawing area when
`clip` is FALSE.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{fatten}{A multiplicative factor used to increase the size of the
middle bar in \code{geom_crossbar()} and the middle point in
\code{geom_pointrange()}.}
}
\value{
a grid grob object.
}
\description{
Convert a ggplot layer to arc grob.
}
\seealso{
Other transform: 
\code{\link{GeomBar2grob}()},
\code{\link{GeomBlank2grob}()},
\code{\link{GeomBoxplot2grob}()},
\code{\link{GeomMarginVtextArcET2grob}()},
\code{\link{GeomPath2grob}()},
\code{\link{GeomPathArcET2grob}()},
\code{\link{GeomPoint2grob}()},
\code{\link{GeomPolygon2grob}()},
\code{\link{GeomPolygonArcET2grob}()},
\code{\link{GeomRaster2grob}()},
\code{\link{GeomRichText2grob}()},
\code{\link{GeomRoundcolArcET2grob}()},
\code{\link{GeomText2grob}()},
\code{\link{GeomViolin2grob}()}
}
\author{
Hou Yun
}
\concept{transform}
