% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer-line.R
\name{GeomPath2grob}
\alias{GeomPath2grob}
\alias{GeomLine2grob}
\alias{GeomSegment2grob}
\alias{GeomSpoke2grob}
\alias{GeomHline2grob}
\alias{GeomVline2grob}
\alias{GeomRug2grob}
\alias{GeomAbline2grob}
\alias{GeomSmooth2grob}
\alias{GeomContour2grob}
\alias{GeomDensity2d2grob}
\alias{GeomQuantile2grob}
\alias{GeomStep2grob}
\alias{GeomFunction2grob}
\title{Convert Layer to Grob}
\usage{
GeomPath2grob(
  data,
  trans = NULL,
  coord = PANEL(),
  region = CELL(),
  ...,
  clip = FALSE,
  na.rm = FALSE
)

GeomLine2grob(
  data,
  trans = NULL,
  coord = PANEL(),
  region = CELL(),
  ...,
  clip = FALSE,
  na.rm = FALSE
)

GeomSegment2grob(
  data,
  trans = NULL,
  coord = PANEL(),
  region = CELL(),
  ...,
  clip = FALSE,
  na.rm = FALSE
)

GeomSpoke2grob(
  data,
  trans = NULL,
  coord = PANEL(),
  region = CELL(),
  ...,
  clip = FALSE,
  na.rm = FALSE
)

GeomHline2grob(
  data,
  trans = NULL,
  coord = PANEL(),
  region = CELL(),
  ...,
  flipped_aes = FALSE,
  clip = FALSE,
  na.rm = FALSE
)

GeomVline2grob(
  data,
  trans = NULL,
  coord = PANEL(),
  region = CELL(),
  ...,
  flipped_aes = FALSE,
  clip = FALSE,
  na.rm = FALSE
)

GeomRug2grob(
  data,
  trans = NULL,
  coord = PANEL(),
  region = CELL(),
  ...,
  sides = "bl",
  clip = FALSE,
  na.rm = FALSE
)

GeomAbline2grob(
  data,
  trans = NULL,
  coord = PANEL(),
  region = CELL(),
  ...,
  flipped_aes = FALSE,
  steps = 0.01,
  simplify = TRUE,
  clip = FALSE,
  na.rm = FALSE
)

GeomSmooth2grob(
  data,
  trans = NULL,
  coord = PANEL(),
  region = CELL(),
  se = TRUE,
  ...,
  flipped_aes = FALSE,
  clip = FALSE,
  na.rm = FALSE
)

GeomContour2grob(
  data,
  trans = NULL,
  coord = PANEL(),
  region = CELL(),
  ...,
  clip = FALSE,
  na.rm = FALSE
)

GeomDensity2d2grob(
  data,
  trans = NULL,
  coord = PANEL(),
  region = CELL(),
  ...,
  clip = FALSE,
  na.rm = FALSE
)

GeomQuantile2grob(
  data,
  trans = NULL,
  coord = PANEL(),
  region = CELL(),
  ...,
  clip = FALSE,
  na.rm = FALSE
)

GeomStep2grob(
  data,
  trans = NULL,
  coord = PANEL(),
  region = CELL(),
  ...,
  flipped_aes = FALSE,
  direction = "hv",
  clip = FALSE,
  na.rm = FALSE
)

GeomFunction2grob(
  data,
  trans = NULL,
  coord = PANEL(),
  region = CELL(),
  ...,
  clip = FALSE,
  na.rm = FALSE
)
}
\arguments{
\item{data}{data frame object, which is extract from a ggplot object.}

\item{trans}{coordinate transform function.}

\item{coord}{coordinate specification, as created by `PANEL()` or extract from
ggplot object.}

\item{region}{a CELL object (created by `CELL()` function) used to set
the drawing area.}

\item{...}{other parameters passing to `Arc*Grob()` function.}

\item{clip}{logical. Allows points to overflow outside the drawing area when
`clip` is FALSE.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{flipped_aes}{TRUE means that coordinates are inherit `CoordFlip`.}

\item{sides}{A string that controls which sides of the plot the rugs appear on.
It can be set to a string containing any of \code{"trbl"}, for top, right,
bottom, and left.}

\item{steps}{step length used to split data. A smaller value means a
smoother curve.}

\item{simplify}{logical, When TRUE, line segments equal to x will not be split.}

\item{se}{Display confidence interval around smooth? (\code{TRUE} by default, see
\code{level} to control.)}

\item{direction}{direction of stairs: 'vh' for vertical then horizontal,
'hv' for horizontal then vertical, or 'mid' for step half-way between
adjacent x-values.}
}
\value{
a grid grob object.
}
\description{
Convert a ggplot layer to arc grob.
}
\seealso{
Other transform: 
\code{\link{GeomBar2grob}()},
\code{\link{GeomBlank2grob}()},
\code{\link{GeomBoxplot2grob}()},
\code{\link{GeomLinerange2grob}()},
\code{\link{GeomMarginVtext2grob}()},
\code{\link{GeomPathArcET2grob}()},
\code{\link{GeomPoint2grob}()},
\code{\link{GeomPolygon2grob}()},
\code{\link{GeomPolygonArcET2grob}()},
\code{\link{GeomRichText2grob}()},
\code{\link{GeomRoundcolArcET2grob}()},
\code{\link{GeomText2grob}()},
\code{\link{GeomViolin2grob}()}
}
\author{
Hou Yun
}
\concept{transform}
