% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer-point.R
\name{GeomPoint2grob}
\alias{GeomPoint2grob}
\alias{GeomHex2grob}
\title{Convert Layer to Grob}
\usage{
GeomPoint2grob(
  data,
  trans = NULL,
  coord = PANEL(),
  region = CELL(),
  ...,
  clip = FALSE,
  na.rm = FALSE
)

GeomHex2grob(
  data,
  trans = NULL,
  coord = PANEL(),
  region = CELL(),
  ...,
  lineend = "butt",
  linejoin = "mitre",
  linemitre = 10,
  clip = FALSE,
  na.rm = FALSE
)
}
\arguments{
\item{data}{data frame object, which is extract from a ggplot object.}

\item{trans}{coordinate transform function.}

\item{...}{other parameters passing to `Arc*Grob()` function.}

\item{clip}{logical. Allows points to overflow outside the drawing area when
`clip` is FALSE.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}
}
\value{
a grid grob object.
}
\description{
Convert a ggplot layer to arc grob.
}
\seealso{
Other transform: 
\code{\link{GeomBar2grob}()},
\code{\link{GeomBoxplot2grob}()},
\code{\link{GeomLinerange2grob}()},
\code{\link{GeomPath2grob}()},
\code{\link{GeomPolygon2grob}()},
\code{\link{GeomViolin2grob}()}
}
\author{
Hou Yun
}
\concept{transform}
