% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer-polygon.R
\name{GeomPolygon2grob}
\alias{GeomPolygon2grob}
\alias{GeomRect2grob}
\alias{GeomTile2grob}
\alias{GeomArea2grob}
\alias{GeomRibbon2grob}
\alias{GeomDensity2grob}
\alias{GeomContourFilled2grob}
\alias{GeomDensity2dFilled2grob}
\title{Convert Layer to Grob}
\usage{
GeomPolygon2grob(
  data,
  trans = NULL,
  coord = PANEL(),
  region = CELL(),
  ...,
  clip = FALSE,
  na.rm = FALSE
)

GeomRect2grob(
  data,
  trans = NULL,
  coord = PANEL(),
  region = CELL(),
  ...,
  clip = FALSE,
  na.rm = FALSE
)

GeomTile2grob(
  data,
  trans = NULL,
  coord = PANEL(),
  region = CELL(),
  ...,
  clip = FALSE,
  na.rm = FALSE
)

GeomArea2grob(
  data,
  trans = NULL,
  coord = PANEL(),
  region = CELL(),
  ...,
  lineend = "butt",
  linejoin = "round",
  linemitre = 10,
  outline.type = "both",
  flipped_aes = FALSE,
  clip = FALSE,
  na.rm = FALSE
)

GeomRibbon2grob(
  data,
  trans = NULL,
  coord = PANEL(),
  region = CELL(),
  ...,
  flipped_aes = FALSE,
  lineend = "butt",
  linejoin = "round",
  linemitre = 10,
  outline.type = "both",
  clip = FALSE,
  na.rm = FALSE
)

GeomDensity2grob(
  data,
  trans = NULL,
  coord = PANEL(),
  region = CELL(),
  ...,
  lineend = "butt",
  linejoin = "round",
  linemitre = 10,
  outline.type = "upper",
  flipped_aes = FALSE,
  clip = FALSE,
  na.rm = FALSE
)

GeomContourFilled2grob(
  data,
  trans = NULL,
  coord = PANEL(),
  region = CELL(),
  ...,
  clip = FALSE,
  na.rm = FALSE
)

GeomDensity2dFilled2grob(
  data,
  trans = NULL,
  coord = PANEL(),
  region = CELL(),
  ...,
  clip = FALSE,
  na.rm = FALSE
)
}
\arguments{
\item{data}{data frame object, which is extract from a ggplot object.}

\item{trans}{coordinate transform function.}

\item{coord}{coordinate specification, as created by `PANEL()` or extract from
ggplot object.}

\item{region}{a CELL object (created by `CELL()` function) used to set
the drawing area.}

\item{...}{other parameters passing to `Arc*Grob()` function.}

\item{clip}{logical. Allows points to overflow outside the drawing area when
`clip` is FALSE.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{lineend}{line end style (round, butt, square).}

\item{linejoin}{line join style (round, mitre, bevel).}

\item{linemitre}{line mitre limit (number greater than 1).}

\item{outline.type}{Type of the outline of the area; \code{"both"} draws both the
upper and lower lines, \code{"upper"}/\code{"lower"} draws the respective lines only.
\code{"full"} draws a closed polygon around the area.}

\item{flipped_aes}{TRUE means that coordinates are inherit `CoordFlip`.}
}
\value{
a grid grob object.
}
\description{
Convert a ggplot layer to arc grob.
}
\seealso{
Other transform: 
\code{\link{GeomBar2grob}()},
\code{\link{GeomBlank2grob}()},
\code{\link{GeomBoxplot2grob}()},
\code{\link{GeomLinerange2grob}()},
\code{\link{GeomMarginVtext2grob}()},
\code{\link{GeomPath2grob}()},
\code{\link{GeomPoint2grob}()},
\code{\link{GeomRichText2grob}()},
\code{\link{GeomRoundcol2grob}()},
\code{\link{GeomText2grob}()},
\code{\link{GeomViolin2grob}()}
}
\author{
Hou Yun
}
\concept{transform}
