% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer-text.R
\name{GeomText2grob}
\alias{GeomText2grob}
\alias{GeomLabel2grob}
\alias{GeomBannerText2grob}
\title{Convert Layer to Grob}
\usage{
GeomText2grob(
  data,
  trans = NULL,
  coord = PANEL(),
  region = CELL(),
  ...,
  parse = FALSE,
  clip = FALSE,
  na.rm = FALSE
)

GeomLabel2grob(
  data,
  trans = NULL,
  coord = PANEL(),
  region = CELL(),
  ...,
  parse = FALSE,
  label.padding = unit(0.25, "lines"),
  label.r = unit(0.15, "lines"),
  label.size = 0.25,
  clip = FALSE,
  na.rm = FALSE
)

GeomBannerText2grob(
  data,
  trans = NULL,
  coord = PANEL(),
  region = CELL(),
  ...,
  clip = FALSE,
  na.rm = FALSE
)
}
\arguments{
\item{data}{data frame object, which is extract from a ggplot object.}

\item{trans}{coordinate transform function.}

\item{coord}{coordinate specification, as created by `PANEL()` or extract from
ggplot object.}

\item{region}{a CELL object (created by `CELL()` function) used to set
the drawing area.}

\item{...}{other parameters passing to `Arc*Grob()` function.}

\item{parse}{If \code{TRUE}, the labels will be parsed into expressions and
displayed as described in \code{?plotmath}.}

\item{clip}{logical. Allows points to overflow outside the drawing area when
`clip` is FALSE.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{label.padding}{Amount of padding around label. Defaults to 0.25 lines.}

\item{label.r}{Radius of rounded corners. Defaults to 0.15 lines.}

\item{label.size}{Size of label border, in mm.}
}
\value{
a grid grob object.
}
\description{
Convert a ggplot layer to arc grob.
}
\seealso{
Other transform: 
\code{\link{GeomBar2grob}()},
\code{\link{GeomBlank2grob}()},
\code{\link{GeomBoxplot2grob}()},
\code{\link{GeomLinerange2grob}()},
\code{\link{GeomMarginVtext2grob}()},
\code{\link{GeomPath2grob}()},
\code{\link{GeomPathArcET2grob}()},
\code{\link{GeomPoint2grob}()},
\code{\link{GeomPolygon2grob}()},
\code{\link{GeomRichText2grob}()},
\code{\link{GeomRoundcolArcET2grob}()},
\code{\link{GeomViolin2grob}()}
}
\author{
Hou Yun
}
\concept{transform}
