% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer-violin.R
\name{GeomViolin2grob}
\alias{GeomViolin2grob}
\title{Convert Layer to Grob}
\usage{
GeomViolin2grob(
  data,
  trans = NULL,
  coord = PANEL(),
  region = CELL(),
  ...,
  flipped_aes = FALSE,
  draw_quantiles = NULL,
  clip = FALSE,
  na.rm = FALSE
)
}
\arguments{
\item{data}{data frame object, which is extract from a ggplot object.}

\item{trans}{coordinate transform function.}

\item{...}{other parameters passing to `Arc*Grob()` function.}

\item{flipped_aes}{TRUE means that coordinates are inherit `CoordFlip`.}

\item{draw_quantiles}{If \code{not(NULL)} (default), draw horizontal lines
at the given quantiles of the density estimate.}

\item{clip}{logical. Allows points to overflow outside the drawing area when
`clip` is FALSE.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}
}
\value{
a grid grob object.
}
\description{
Convert a ggplot layer to arc grob.
}
\seealso{
Other transform: 
\code{\link{GeomBar2grob}()},
\code{\link{GeomBoxplot2grob}()},
\code{\link{GeomLinerange2grob}()},
\code{\link{GeomPath2grob}()},
\code{\link{GeomPoint2grob}()},
\code{\link{GeomPolygon2grob}()}
}
\author{
Hou Yun
}
\concept{transform}
