% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcET-global.R
\name{cell_id}
\alias{cell_id}
\alias{track_id}
\alias{sector_id}
\alias{reset_ids}
\title{Get ids}
\usage{
cell_id()

track_id()

sector_id()

reset_ids(which = c("all", "CellID", "TrackID", "SectorID"))
}
\arguments{
\item{which}{if "all" will reset all id's index.}
}
\value{
`*_id()` return a character vector.
}
\description{
Helper functions used to generate "CellID", "TrackID" and "SectorID".
}
\examples{
cell_id()
## increment 1
cell_id()

## reset
reset_ids()
cell_id()
}
\author{
Hou Yun
}
