% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc.R
\name{init_cell}
\alias{init_cell}
\title{Add CellPlot}
\usage{
init_cell(
  plot,
  data = NULL,
  mapping = aes(),
  region = CELL(),
  CellID = NULL,
  TrackID = NULL,
  SectorID = NULL,
  inherit.aes = TRUE
)
}
\arguments{
\item{plot}{an ArcPlot object.}

\item{data}{default dataset to use for this CellPlot. If not specified, must be
supplied in each layer. Note: data can also be an ggplot object.}

\item{mapping}{default list of aesthetic mappings to use for plot. If not
specified, must be supplied in each layer.}

\item{region}{a CELL object (created by `CELL()` function) used to set
the drawing area.}

\item{CellID, TrackID, SectorID}{ids specification, must be one-length character.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}
}
\value{
an modified ArcPlot object.
}
\description{
`init_cell()` add a new CellPlot on ArcPlot.
}
\author{
Hou Yun
}
