% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer-bar.R
\name{layer_bar}
\alias{layer_bar}
\alias{layer_col}
\alias{layer_histogram}
\title{Bar}
\usage{
layer_bar(..., show.legend = FALSE)

layer_col(..., show.legend = FALSE)

layer_histogram(..., show.legend = FALSE)
}
\arguments{
\item{...}{extra parameters passing to `geom_*()` function.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}
}
\value{
a gg layer object.
}
\description{
These functions are an encapsulation of relative functions in
the ggplot2 package, and the only difference is that these functions do
not add legends by default.
}
\seealso{
Other layer: 
\code{\link{layer_blank}()},
\code{\link{layer_boxplot}()},
\code{\link{layer_path}()},
\code{\link{layer_pointrange}()},
\code{\link{layer_point}()},
\code{\link{layer_polygon}()},
\code{\link{layer_text}()},
\code{\link{layer_violin}()}
}
\author{
Hou Yun
}
\concept{layer}
