% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer-line.R
\name{layer_path}
\alias{layer_path}
\alias{layer_line}
\alias{layer_segment}
\alias{layer_hline}
\alias{layer_vline}
\alias{layer_rug}
\alias{layer_abline}
\alias{layer_spoke}
\alias{layer_smooth}
\alias{layer_step}
\alias{layer_contour}
\alias{layer_density_2d}
\alias{layer_density2d}
\alias{layer_qq_line}
\alias{layer_quantile}
\alias{layer_function}
\title{Path Layer}
\usage{
layer_path(..., show.legend = FALSE)

layer_line(..., show.legend = FALSE)

layer_segment(..., show.legend = FALSE)

layer_hline(..., show.legend = FALSE)

layer_vline(..., show.legend = FALSE)

layer_rug(..., show.legend = FALSE)

layer_abline(..., show.legend = FALSE)

layer_spoke(..., show.legend = FALSE)

layer_smooth(..., show.legend = FALSE)

layer_step(..., show.legend = FALSE)

layer_contour(..., show.legend = FALSE)

layer_density_2d(..., show.legend = FALSE)

layer_density2d(..., show.legend = FALSE)

layer_qq_line(..., show.legend = FALSE)

layer_quantile(..., show.legend = FALSE)

layer_function(..., show.legend = FALSE)
}
\arguments{
\item{...}{extra parameters passing to `geom_*()` function.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}
}
\value{
a gg layer object.
}
\description{
These functions are an encapsulation of relative functions in
the ggplot2 package, and the only difference is that these functions do
not add legends by default.
}
\seealso{
Other layer: 
\code{\link{layer_bar}()},
\code{\link{layer_blank}()},
\code{\link{layer_boxplot}()},
\code{\link{layer_pointrange}()},
\code{\link{layer_point}()},
\code{\link{layer_polygon}()},
\code{\link{layer_text}()},
\code{\link{layer_violin}()}
}
\author{
Hou Yun
}
\concept{layer}
