% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer-polygon.R
\name{layer_polygon}
\alias{layer_polygon}
\alias{layer_rect}
\alias{layer_tile}
\alias{layer_area}
\alias{layer_contour_filled}
\alias{layer_density}
\alias{layer_density2d_filled}
\alias{layer_density_2d_filled}
\alias{layer_freqpoly}
\title{Polygon Layer}
\usage{
layer_polygon(..., show.legend = FALSE)

layer_rect(..., show.legend = FALSE)

layer_tile(..., show.legend = FALSE)

layer_area(..., show.legend = FALSE)

layer_contour_filled(..., show.legend = FALSE)

layer_density(..., show.legend = FALSE)

layer_density2d_filled(..., show.legend = FALSE)

layer_density_2d_filled(..., show.legend = FALSE)

layer_freqpoly(..., show.legend = FALSE)
}
\arguments{
\item{...}{extra parameters passing to `geom_*()` function.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}
}
\value{
a gg layer object.
}
\description{
These functions are an encapsulation of relative functions in
the ggplot2 package, and the only difference is that these functions do
not add legends by default.
}
\seealso{
Other layer: 
\code{\link{layer_bar}()},
\code{\link{layer_blank}()},
\code{\link{layer_boxplot}()},
\code{\link{layer_path}()},
\code{\link{layer_pointrange}()},
\code{\link{layer_point}()},
\code{\link{layer_violin}()}
}
\author{
Hou Yun
}
\concept{layer}
