% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{map_colour}
\alias{map_colour}
\alias{map_color}
\alias{map_colour_manual}
\alias{map_color_manaul}
\alias{map_fill}
\alias{map_fill_manual}
\alias{map_shape}
\alias{map_shape_manual}
\alias{map_linetype}
\alias{map_linetype_manual}
\alias{map_size}
\alias{map_size_manual}
\alias{map_linewidth}
\alias{map_linewidth_manual}
\title{Mapping Functions}
\usage{
map_colour(x, palette = NULL, domain = NULL, ...)

map_color(x, palette = NULL, domain = NULL, ...)

map_colour_manual(x, values = NULL, na.value = "grey60", ...)

map_color_manaul(x, values = NULL, na.value = "grey60", ...)

map_fill(x, palette = NULL, domain = NULL, ...)

map_fill_manual(x, values = NULL, na.value = "grey60", ...)

map_shape(x, ...)

map_shape_manual(x, values = NULL, na.value = NA, ...)

map_linetype(x, ...)

map_linetype_manual(x, values = NULL, na.value = NA, ...)

map_size(x, range = c(1, 6), ...)

map_size_manual(x, values = NULL, na.value = NA, ...)

map_linewidth(x, range = c(0.1, 1), ...)

map_linewidth_manual(x, values = NULL, na.value = NA, ...)
}
\arguments{
\item{x}{numeric or character vector, which will be mapped to aesthetic.}

\item{palette}{The colours or colour function that values will be mapped to}

\item{domain}{The possible values that can be mapped.

For \code{col_numeric} and \code{col_bin}, this can be a simple numeric
range (e.g. \code{c(0, 100)}); \code{col_quantile} needs representative
numeric data; and \code{col_factor} needs categorical data.

If \code{NULL}, then whenever the resulting colour function is called, the
\code{x} value will represent the domain. This implies that if the function
is invoked multiple times, the encoding between values and colours may not
be consistent; if consistency is needed, you must provide a non-\code{NULL}
domain.}

\item{...}{other parameters.}

\item{values}{a set of aesthetic values to map data values to. The values
will be matched in order (usually alphabetical) with the limits of the
scale, or with \code{breaks} if provided. If this is a named vector, then the
values will be matched based on the names instead. Data values that don't
match will be given \code{na.value}.}

\item{na.value}{The aesthetic value to use for missing (\code{NA}) values}

\item{range}{Numeric vector of length two, giving range of possible sizes.
Should be greater than 0.}
}
\value{
a mapped_aes vector.
}
\description{
Helper Functions for Mapping.
}
\author{
Hou Yun
}
