% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{PANEL}
\alias{PANEL}
\title{Panel Meta Information}
\usage{
PANEL(
  x_limits = NULL,
  y_limits = NULL,
  x_breaks = NULL,
  y_breaks = NULL,
  x_minor_breaks = NULL,
  y_minor_breaks = NULL,
  x_labels = NULL,
  y_labels = NULL,
  x_range = NULL,
  y_range = NULL,
  x_expand = NULL,
  y_expand = NULL,
  drop = FALSE
)
}
\arguments{
\item{x_limits, y_limits}{For continuous coordinates, it is a numerical
vector of length 2, representing the value range of coordinate axes.
For discrete coordinates, it is an enumeration vector of all
that can take values.}

\item{x_breaks, y_breaks}{For continuous coordinates, it is a numerical
vector in the interval, representing the position of the ticks.
For discrete coordinates, it's anything within the limits.}

\item{x_labels, y_labels}{labels of ticks, should be same length as breaks.}

\item{x_expand, y_expand}{a vector of range expansion constants used to add
some padding around the data to ensure that they are placed some distance
away from the axes. It can be created by `ggplot2::expansion()`.}

\item{drop}{logical. If TRUE, the levels that does not appear will be removed.}
}
\value{
a `PANEL` object.
}
\description{
This function always used to setting panel meta information on
cartesian coordinate.
the end point.
}
\author{
Hou Yun
}
