% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcET-global.R
\name{set_arcET_global}
\alias{set_arcET_global}
\alias{get_arcET_global}
\title{Get and set global options}
\usage{
set_arcET_global(..., env = NULL)

get_arcET_global(which, default = NULL, env = NULL)
}
\arguments{
\item{...}{any options can be defined, using name = value. Currently,
only the `cell_index`,`sector_index`,`track_index`,`steps` and `simplify`
options are used}

\item{env}{the environment in which IDs is to be generated. If is NULL (default), will
evaluated in `arcET_global` environment.}

\item{which}{a character string indicating which global setting will be returned.}

\item{default}{if this global setting is not setting, the default will be returned.}
}
\value{
for `set_arcET_global()`, will return invisible `NULL`. And for `get_arcET_global()`,
will return global setting or default.
}
\description{
Helper functions used to get and set global options.
}
\author{
Hou Yun
}
