% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-cell.R
\name{show_cell}
\alias{show_cell}
\title{Show CELL}
\usage{
show_cell(
  ...,
  fill = "lightblue",
  colour = NA,
  background = "grey95",
  border = NA,
  digits = 2,
  xn = 7,
  yn = 7,
  newpage = is.null(vp),
  vp = NULL
)
}
\arguments{
\item{...}{any CELL objects (created by `CELL()` function) used to set
the drawing area.}

\item{fill}{fill colour of region.}

\item{colour}{border colour of region.}

\item{background, border}{fill and border colour of plot region (unit circle).}

\item{digits}{integer indicating the number of decimal places.}

\item{xn, yn}{number of xaxis/yaxis ticks.}

\item{newpage}{logical, indicating whether a new graphics device needs to be opened.}

\item{vp}{grid viewport object.}
}
\value{
invisible grobs.
}
\description{
A quick way to show region of `CELL()`.
}
\examples{
show_cell(CELL(120, 60, 0.6))
show_cell(CELL(120, 60, 0.6), CELL(300, 240, 0.3, 0.6))
}
\author{
Hou Yun
}
