% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc-test.R
\name{arc_test}
\alias{arc_test}
\title{Test function}
\usage{
arc_test(
  plot = ggplot2::last_plot(),
  region = NULL,
  ...,
  vp = NULL,
  newpage = is.null(vp)
)
}
\arguments{
\item{plot}{a ggplot object.}

\item{region}{a `CELL` object.}

\item{...}{passing to `within_plot()`.}

\item{vp}{viewport to draw plot in.}

\item{newpage}{draw new (empty) page first?}
}
\value{
return grob object invisibly.
}
\description{
Helper funtion used to quickly preview the results of the conversion.
}
\examples{
if (FALSE) { # Not Run
  library(ggplot2)
  library(arcET)
  ggplot(mtcars, aes(wt, mpg)) + geom_point()
  arc_test()

  ggplot(mpg, aes(class)) + geom_bar()
  arc_test()
}
}
\author{
Hou Yun
}
