% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc.R
\name{arcplot}
\alias{arcplot}
\title{Create a new ArcPlot}
\usage{
arcplot(data = NULL, mapping = aes(), ...)
}
\arguments{
\item{data}{default dataset to use for arcplot. If not specified, must be
supplied in each cell plot.}

\item{mapping}{default list of aesthetic mappings to use for plot. If not
specified, must be supplied in each cell plot.}

\item{...}{not used.}
}
\value{
an ArcPlot object.
}
\description{
`arcplot()` initializes a ArcPlot object, and It works similarly
to `ggplot()`.
}
\author{
Hou Yun
}
