% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{ArcPlot_build}
\alias{ArcPlot_build}
\alias{ArcPlot_build.ArcPlot}
\title{Build ArcPlot}
\usage{
ArcPlot_build(plot, ...)

\method{ArcPlot_build}{ArcPlot}(
  plot,
  theme = NULL,
  title = NULL,
  subtitle = NULL,
  tag = NULL,
  caption = NULL,
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{plot}{an ArcPlot object.}

\item{...}{other parameters passing to `merge_guides()`.}

\item{theme}{a ggplot theme object.}

\item{title}{plot title.}

\item{subtitle}{plot subtitle.}

\item{tag}{plot tag.}

\item{caption}{plot caption.}

\item{xlim, ylim}{two-length numeric vector used to set scale limits.}
}
\value{
grid grobs.
}
\description{
Internal function used to convert ArcPlot to grobs.
}
\author{
Hou Yun
}
