% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last.R
\name{arcsave}
\alias{arcsave}
\title{Export ArcPlot}
\usage{
arcsave(filename, plot = current_plot(), ...)
}
\arguments{
\item{filename}{file name to create on disk.}

\item{plot}{an ArcPlot object.}

\item{...}{other parameters passing to `ggsave()`.}
}
\description{
This function is modified (copy?) from `ggplot2::ggsave`, and used
to export ArcPlot object.
}
\keyword{internal}
