% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.R
\name{ArcxAxisGrob}
\alias{ArcxAxisGrob}
\title{Arc x-axis Grob}
\usage{
ArcxAxisGrob(
  coord = PANEL(),
  region = CELL(),
  position = "top",
  ticks.length = 1.5,
  line.gp = element_line(),
  tick.gp = element_line(),
  text.gp = element_text(),
  steps = 0.01,
  ...
)
}
\arguments{
\item{region}{a CELL object (created by `CELL()` function) used to set
the drawing area.}

\item{ticks.length}{a grid unit object or numeric, and numerical value
means that the length of ticks is measured in millimeters.}

\item{line.gp}{should be created by `ggplot2::element_line()`.}

\item{text.gp}{should be created by `ggplot2::element_text()`.}

\item{steps}{step length used to split data. A smaller value means a
smoother curve.}

\item{...}{not used.}

\item{title}{character used to specifying x-axis's title.}

\item{breaks}{numeric vector of x-value [in degree] locations for the tick marks.}

\item{labels}{character vector of tick's label.}

\item{postion}{"bottom" or "top" used to specifying axis position,
and default is "bottom"}

\item{facing}{"clockwise" or "banner".}

\item{title.gp}{should be created by `ggplot2::element_text()`.}

\item{ticks.gp}{should be created by `ggplot2::element_line()`.}
}
\value{
a `xAxisGrob` object.
}
\description{
These functions can draw x-axis on polar coordinate.
}
\author{
Hou Yun
}
