% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-cell.R
\name{show_cell}
\alias{show_cell}
\title{Show CELL}
\usage{
show_cell(
  region = CELL(),
  fill = "grey90",
  colour = NA,
  digits = 2,
  xn = 7,
  yn = 7,
  newpage = is.null(vp),
  vp = NULL,
  ...
)
}
\arguments{
\item{region}{a CELL object (created by `CELL()` function) used to set
the drawing area.}

\item{fill}{fill colour of region.}

\item{colour}{border colour of region.}

\item{digits}{integer indicating the number of decimal places.}

\item{xn, yn}{number of xaxis/yaxis ticks.}

\item{newpage}{logical, indicating whether a new graphics device needs to be opened.}

\item{vp}{grid viewport object.}

\item{...}{other parameters passing to `ArcRectGrob()`.}
}
\value{
invisible grobs.
}
\description{
A quick way to show region of `CELL()`.
}
\examples{
show_cell(CELL(start = 120, end = 60, r0 = 0.5))
}
\author{
Hou Yun
}
