% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/references.R
\name{within_cell}
\alias{within_cell}
\alias{within_track}
\alias{within_sector}
\alias{within_plot}
\title{Modify CellPlot}
\usage{
within_cell(plot, ..., CellID = NULL)

within_track(plot, ..., TrackID = NULL)

within_sector(plot, ..., SectorID = NULL)

within_plot(plot, ...)
}
\arguments{
\item{plot}{an ArcPlot object.}

\item{...}{elements will be added.}

\item{CellID, TrackID, SectorID}{character IDs indicating which element will be
modified.}
}
\value{
an modified ArcPlot object.
}
\description{
Helper functions to modify ArcPlot elements by IDs.
}
\author{
Hou Yun
}
